Persist Fish

+++ SpaceCat/Assets/Scripts/Stage Controls/GroundRepeat.cs
@@ -10,13 +10,11 @@ public class GroundRepeat : MonoBehaviour {
 	
 	private float screenWidth;
 
-	// Use this for initialization
 	void Start () {
 		float screenHeight = 2.0f * Camera.main.orthographicSize;
 		screenWidth = screenHeight * Camera.main.aspect;
 	}
-	
-	// Update is called once per frame
+
 	void Update () {
 		GenerateGround ();
 	}
@@ -27,7 +25,7 @@ public class GroundRepeat : MonoBehaviour {
 		GameObject ground = (GameObject)Instantiate(availableGround);
 		float groundWidth = ground.transform.localScale.x;
 		float groundCenter = groundEndX + groundWidth * 0.5f;
-		ground.transform.position = new Vector3((float)(groundEndX + 13.85), (float)(-3.7), 0);
+		ground.transform.position = new Vector3((float)(groundEndX + 13.85), -3.7f, 0);
 		currentGround.Add(ground);         
 	}
 
@@ -43,31 +41,25 @@ public class GroundRepeat : MonoBehaviour {
 		
 		foreach(var ground in currentGround)
 		{
-			//7
 			float groundWidth = ground.transform.localScale.x;
 			float groundStartX = ground.transform.position.x - (groundWidth * 0.5f);    
 			float groundEndX = groundStartX + groundWidth;                            
-			
-			//8
+
 			if (groundStartX > addGroundX)
 				addGround = false;
-			
-			//9
+
 			if (groundEndX < removeGroundX)
 				groundToRemove.Add(ground);
-			
-			//10
+
 			farGroundEndX = Mathf.Max(farGroundEndX, groundEndX);
 		}
-		
-		//11
+
 		foreach(var ground in groundToRemove)
 		{
 			currentGround.Remove(ground);
 			Destroy(ground);            
 		}
-		
-		//12
+
 		if (addGround)
 			AddGround(farGroundEndX);
 	}
